// src/frontend/web/react-admin/src/generated/models/sanction-prevu.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mon API
 * Documentation de l'API de votre application.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface SanctionPrevu
 */
export interface SanctionPrevu {
    /**
     * 
     * @type {number}
     * @memberof SanctionPrevu
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof SanctionPrevu
     */
    libelle: any;
    /**
     * 
     * @type {string}
     * @memberof SanctionPrevu
     */
    niveauGravite?: any;
    /**
     * 
     * @type {string}
     * @memberof SanctionPrevu
     */
    motif: any;
    /**
     * 
     * @type {string}
     * @memberof SanctionPrevu
     */
    dureeValidite: any;
    /**
     * 
     * @type {Eleve}
     * @memberof SanctionPrevu
     */
    eleve?: any;
    /**
     * 
     * @type {Convocation}
     * @memberof SanctionPrevu
     */
    convocation?: any;
    /**
     * 
     * @type {Regle}
     * @memberof SanctionPrevu
     */
    regle?: any;
    /**
     * 
     * @type {Faute}
     * @memberof SanctionPrevu
     */
    faute?: any;
    /**
     * 
     * @type {BaseModelpropertiescreatedAt}
     * @memberof SanctionPrevu
     */
    createdAt?: any;
    /**
     * 
     * @type {BaseModelpropertiesupdatedAt}
     * @memberof SanctionPrevu
     */
    updatedAt?: any;
    /**
     * 
     * @type {BaseModelpropertiesdeletedAt}
     * @memberof SanctionPrevu
     */
    deletedAt?: any;
}